Forgot to add test cases for the different cases the letters could be in; I just tested it within running the calculator. And changed the answer type to duouble to allow for bigger numbers to be represented, but I don't think that worked...

+++ coursework2/src/main/java/GuiView.java
@@ -32,8 +32,8 @@ public class GuiView {
 
   @FXML
   void isPressed(ActionEvent event) {
-    float answer = calc.evaluate(getExpression());
-    String displayString = Float.toString(answer);
+    double answer = calc.evaluate(getExpression());
+    String displayString = Double.toString(answer);
     displayAnswer.setText(displayString);
   }
